feat: add a matrix inversion backend using Eigen3
